refactor: shared
