+++ packages/compat/plugin-swc/src/utils.ts
@@ -1,12 +1,11 @@
 import fs from 'node:fs';
+import path from 'node:path';
 import { __internalHelper } from '@rsbuild/core';
 import {
   type ModifyChainUtils,
   type NormalizedConfig,
-  findUp,
   getBrowserslistWithDefault,
   getCoreJsVersion,
-  getDefaultStyledComponentsConfig,
   isUsingHMR,
 } from '@rsbuild/shared';
 import semver from '@rsbuild/shared/semver';
@@ -20,6 +19,30 @@ import type {
 } from './types';
 
 const { applySwcDecoratorConfig } = __internalHelper;
+
+async function findUp({
+  filename,
+  cwd = process.cwd(),
+}: {
+  filename: string;
+  cwd?: string;
+}) {
+  const { root } = path.parse(cwd);
+
+  let dir = cwd;
+  while (dir && dir !== root) {
+    const filePath = path.join(dir, filename);
+
+    try {
+      const stats = await fs.promises.stat(filePath);
+      if (stats?.isFile()) {
+        return filePath;
+      }
+    } catch {}
+
+    dir = path.dirname(dir);
+  }
+}
 
 const isBeyondReact17 = async (cwd: string) => {
   const pkgPath = await findUp({ cwd, filename: 'package.json' });
@@ -141,6 +164,17 @@ export async function finalizeConfig(
 
   return finalized;
 }
+
+const getDefaultStyledComponentsConfig = (isProd: boolean, ssr: boolean) => {
+  return {
+    ssr,
+    // "pure" is used to improve dead code elimination in production.
+    // we don't need to enable it in development because it will slow down the build process.
+    pure: isProd,
+    displayName: true,
+    transpileTemplateLiterals: true,
+  };
+};
 
 export async function applyPluginConfig(
   rawOptions: PluginSwcOptions,
