+++ packages/compat/webpack/src/shared.ts
@@ -1,10 +1,4 @@
-import fs from 'node:fs';
-import { join } from 'node:path';
 import { __internalHelper } from '@rsbuild/core';
-import {
-  type SharedCompiledPkgNames,
-  getSharedPkgCompiledPath,
-} from '@rsbuild/shared';
 
 const {
   plugins,
@@ -37,11 +31,3 @@ export {
 };
 
 export type InternalContext = __internalHelper.InternalContext;
-
-export const getCompiledPath = (packageName: string) => {
-  const providerCompilerPath = join(__dirname, '../../compiled', packageName);
-  if (fs.existsSync(providerCompilerPath)) {
-    return providerCompilerPath;
-  }
-  return getSharedPkgCompiledPath(packageName as SharedCompiledPkgNames);
-};
