+++ packages/core/src/loadEnv.ts
@@ -1,8 +1,9 @@
 import fs from 'node:fs';
 import { join } from 'node:path';
-import { getNodeEnv, isFileSync } from '@rsbuild/shared';
+import { getNodeEnv } from '@rsbuild/shared';
 import { expand } from '../compiled/dotenv-expand/index.js';
 import { parse } from '../compiled/dotenv/index.js';
+import { isFileSync } from './helpers';
 
 export type LoadEnvOptions = {
   /**
