+++ packages/core/src/plugins/less.ts
@@ -9,7 +9,7 @@ import {
   getSharedPkgCompiledPath,
   mergeChainedOptions,
 } from '@rsbuild/shared';
-import { getCompiledPath } from '../provider/shared';
+import { getCompiledPath } from '../helpers';
 import type { RsbuildPlugin } from '../types';
 
 const getLessLoaderOptions = (
