+++ packages/core/tests/cache.test.ts
@@ -1,7 +1,7 @@
 import { createStubRsbuild } from '@scripts/test-helper';
 import { pluginCache } from '../src/plugins/cache';
 
-vi.mock('@rsbuild/shared', async (importOriginal) => {
+vi.mock('../src/helpers.js', async (importOriginal) => {
   const mod = await importOriginal<any>();
   return {
     ...mod,
