+++ packages/core/tests/swc.test.ts
@@ -1,7 +1,7 @@
 import type { RsbuildConfig } from '@rsbuild/shared';
 import { createStubRsbuild } from '@scripts/test-helper';
 import { pluginEntry } from '../src/plugins/entry';
-import { pluginSwc } from '../src/provider/plugins/swc';
+import { pluginSwc } from '../src/plugins/swc';
 
 describe('plugin-swc', () => {
   it('should disable preset_env in target other than web', async () => {
