+++ packages/shared/src/index.ts
@@ -7,5 +7,4 @@ export * from './mergeChainedOptions';
 export * from './types';
 export * from './config';
 export * from './chain';
-export * from './minimize';
 export * from './url';
