+++ packages/shared/src/utils.ts
@@ -71,17 +71,6 @@ export interface AwaitableGetter<T> extends PromiseLike<T[]> {
   promises: Promise<T>[];
 }
 
-/**
- * Make Awaitable.
- */
-export const awaitableGetter = <T>(
-  promises: Promise<T>[],
-): AwaitableGetter<T> => {
-  const then: PromiseLike<T[]>['then'] = (...args) =>
-    Promise.all(promises).then(...args);
-  return { then, promises };
-};
-
 export const getJsSourceMap = (config: NormalizedConfig) => {
   const { sourceMap } = config.output;
   if (sourceMap.js === undefined) {
@@ -109,13 +98,6 @@ export function isServerTarget(target: RsbuildTarget[]) {
   );
 }
 
-export function resolvePackage(loader: string, dirname: string) {
-  // Vitest do not support require.resolve to source file
-  return process.env.VITEST
-    ? loader
-    : require.resolve(loader, { paths: [dirname] });
-}
-
 export const getCoreJsVersion = (corejsPkgPath: string) => {
   try {
     const { version } = fse.readJSONSync(corejsPkgPath);
@@ -125,15 +107,6 @@ export const getCoreJsVersion = (corejsPkgPath: string) => {
     return '3';
   }
 };
-
-/**
- * ensure absolute file path.
- * @param base - Base path to resolve relative from.
- * @param filePath - Absolute or relative file path.
- * @returns Resolved absolute file path.
- */
-export const ensureAbsolutePath = (base: string, filePath: string): string =>
-  path.isAbsolute(filePath) ? filePath : path.resolve(base, filePath);
 
 export const castArray = <T>(arr?: T | T[]): T[] => {
   if (arr === undefined) {
