+++ packages/shared/tests/utils.test.ts
@@ -1,14 +1,4 @@
-import { awaitableGetter, camelCase, createDependenciesRegExp } from '../src';
-
-describe('awaitableGetter', () => {
-  it('should work', async () => {
-    const arr = ['123', '456'];
-    const promises = arr.map((item) => Promise.resolve(item));
-    const wrapped = awaitableGetter(promises);
-    expect(wrapped.promises).toStrictEqual(promises);
-    expect(await wrapped).toStrictEqual(arr);
-  });
-});
+import { camelCase, createDependenciesRegExp } from '../src';
 
 describe('camelCase', () => {
   test('should convert snake_case to camelCase', () => {
